+++ dev_test_cex_full_non_stop.py
@@ -63,8 +63,9 @@ print("starting monitoring api!")
 binance_websocket_api_manager.start_monitoring_api()
 
 # set api key and secret for userData stream
-binance_api_key = ""
-binance_api_secret = ""
+binance_api_key = "OMfMlElvlYtwoaHIRtBjuUZozKddKoap1Z1qXESaZVX9pNQq396RqARRR9FOhdaq"
+binance_api_secret = "XHllFpxy3Vd7RRTNl1Yhpo7PsmJeM3MICPNFpGYOhnFK8ZjcOAY1045DBb5udMOe"
+
 binance_websocket_api_manager.set_private_api_config(binance_api_key, binance_api_secret)
 userdata_stream_id = binance_websocket_api_manager.create_stream(["arr"], ["!userData"])
 
@@ -100,11 +101,7 @@ binance_websocket_api_manager.create_stream(["depth10"], markets)
 binance_websocket_api_manager.create_stream(["depth20"], markets)
 binance_websocket_api_manager.create_stream(["aggTrade"], markets)
 
-markets = {'bnbbtc', 'ethbtc', 'btcusdt', 'bchabcusdt', 'xrpusdt', 'rvnbtc', 'ltcusdt', 'adausdt', 'eosusdt',
-           'neobtc', 'adaeth', 'icxusdt', 'btctusd', 'icxbtc', 'btcusdc', 'wanbtc', 'zecbtc', 'wtcbtc',
-           'batbtc', 'adabnb', 'etcusdt', 'qtumusdt', 'xmrbtc', 'trxeth', 'adatusd', 'trxxrp', 'trxbnb',
-           'ltctusd', 'polybnb', 'scbtc', 'steembtc', 'trxtusd', 'npxseth', 'kmdbtc', 'polybtc', 'gasbtc',
-           'bnbpax', 'linkusdt', 'hceth', 'zrxeth', 'icxeth', 'xmreth', 'neobnb', 'etceth', 'zeceth', 'xmrbnb'}
+
 channels = {'trade', 'kline_1m', 'kline_5m', 'kline_15m', 'kline_30m', 'kline_1h', 'kline_12h', 'kline_1w',
             'miniTicker', 'depth20'}
 id = binance_websocket_api_manager.create_stream(channels, markets)