+++ example_stream_buffer.py
@@ -62,8 +62,11 @@ markets = {'bnbbtc', 'ethbtc', 'btcusdt', 'bchabcusdt', 'xrpusdt', 'rvnbtc', 'lt
            'bnbpax', 'linkusdt', 'hceth', 'zrxeth', 'icxeth', 'xmreth', 'neobnb', 'etceth', 'zeceth', 'xmrbnb',
            'wanbnb', 'zrxbnb', 'agibnb', 'funeth', 'arketh', 'engeth'}
 
-binance_get_kline_stream_id1 = binance_websocket_api_manager.create_stream(['kline_1m', 'kline_5m'], markets)
-binance_get_kline_stream_id2 = binance_websocket_api_manager.create_stream(['kline_30m', 'kline_1h', 'kline_15m'], markets)
+channels = ['kline_1m', 'kline_5m', 'kline_15m', 'kline_30m', 'kline_1h', 'kline_12h', 'kline_1w']
+binance_get_kline_stream_id1 = binance_websocket_api_manager.create_stream(channels, markets)
+
+channels = {'trade', 'miniTicker', 'depth20'}
+binance_get_kline_stream_id2 = binance_websocket_api_manager.create_stream(channels=channels, markets=markets)
 
 
 def print_stream_data_from_stream_buffer(binance_websocket_api_manager):